Gemfile: Lock nokogiri to Ruby 2.6-compatible version
This paves the way for https://github.com/perlun/perlun.eu.org/pull/40.
We could outright upgrade Nokogiri in this PR, but I'll let Dependabot
take the credit for that since it was the one who prompted the upgrade
in the first place. :)

The problem with https://github.com/perlun/perlun.eu.org/pull/40 as it
currently stands is that it tries to use Nokogiri 1.16.0, which has
dropped both Ruby 2.6 support (in 1.14.0) and Ruby 2.7 (in 1.16.0). Both
of these Ruby versions have reached EOL, so we should certainly upgrade
our Ruby version as well but let's begin with addressing these
security-related PRs first.